Merge app.py from dev
This enables two demo scenarios; local and live.
This change is needed because huggingface spaces does not
have access to the webcam of the browser's host machine.

This commit also prepares for the YOLO model requirements by
adding the ultralytics library, since we will be using the `.pt`
file rather than an ONNX file because of time constraints